requirements_pip.txt


flask
flask_session
wechatpayv3
SQLAlchemy==1.4.46
sqlparse
termcolor
PyPDF2

